Consistency refactor (pass 1): GaffTapeAlign, LimelightSeek, LimelightSmoothTurn, PuppyDog

+++ src/main/java/frc/robot/lib/vision/Limelight.java
@@ -201,8 +201,8 @@ public class Limelight {
      * True active pipeline index of the camera.
      * @return 0 - 9
      */
-    public double getPipe() {
-        return getDouble("getpipe");
+    public int getPipe() {
+        return (int) getDouble("getpipe");
     }
 
     /**
